refactor: rename mainView to ConferenceMainView as that is what the component is
