+++ src/components/conference.jsx
@@ -10,7 +10,7 @@ import {
   useHMSStore,
 } from "@100mslive/react-sdk";
 import { Box, Flex } from "@100mslive/roomkit-react";
-import { ConferenceMainView } from "../layouts/mainView";
+import { ConferenceMainView } from "../layouts/ConferenceMainView";
 import { Footer } from "./Footer";
 import FullPageProgress from "./FullPageProgress";
 import { Header } from "./Header";
