+++ src/layouts/ConferenceMainView.jsx
